[修正] ディレクトリの整理とDrawMenuの実装 - ディレクトリをDriverとlibに分割 - mainに処理内容を書かないようにDrawMenuを実装
