+++ src/main.cpp
@@ -8,49 +8,12 @@
 #include <Adafruit_TinyUSB.h>
 #include <MIDI.h>
 
-#include "RotaryEncoder.h"
-#include "PedalMonitor.h"
-#include "ExpressionPedal.h"
+#include "config.h"
 
-#define EXP_PIN 28
-
-#define CLK_PIN 15
-#define DT_PIN 26
-#define RE_SW_PIN 27
-
-#define SCL1_PIN 3
-#define SDA1_PIN 2
-
-#define SCL0_PIN 5
-#define SDA0_PIN 4
-
-#define PEDAL1_PIN 13
-#define PEDAL2_PIN 14
-#define PEDAL3_PIN 10
-#define PEDAL4_PIN 9
-#define PEDAL5_PIN 12
-#define PEDAL6_PIN 11
-
-#define WS2812_PIN 16
-
-#define SCREEN_WIDTH 128
-#define SCREEN_HEIGHT 64
-
-#define FONT_WIDTH 8
-#define FONT_HEIGHT 6
-
-#define OLED_RESET     -1
-#define SCREEN_ADDRESS 0x3C
-
-#define PEDAL_NUMBER 6
-
-#define EXP_CH 0
-
-const int PEDAL_PIN[] = {PEDAL1_PIN, PEDAL2_PIN, PEDAL3_PIN, PEDAL4_PIN, PEDAL5_PIN, PEDAL6_PIN};
-
-int PedalControlNumber[] = {1,2,3,4,5,6};
-
-Adafruit_SSD1306 display(SCREEN_WIDTH, SCREEN_HEIGHT, &Wire, OLED_RESET);
+#include "driver/RotaryEncoder/RotaryEncoder.h"
+#include "driver/PedalMon/PedalMon.h"
+#include "driver/ExpPedal/ExpPedal.h"
+#include "driver/DrawMenu/DrawMenu.h"
 
 Adafruit_NeoPixel pixels(1, WS2812_PIN, NEO_GRB + NEO_KHZ800);
 
@@ -60,9 +23,11 @@ Adafruit_USBD_MIDI usb_midi;
 
 MIDI_CREATE_INSTANCE(Adafruit_USBD_MIDI, usb_midi, MIDI);
 
-ExpressionPedal expPedal(&MIDI, 0);
+ExpPedal expPedal(&MIDI, 0);
 
-PedalMonitor myPedals(&MIDI, PEDAL_PIN, PedalControlNumber, PEDAL_NUMBER);
+PedalMon myPedals(&MIDI, PEDAL_PIN, PEDAL_CON_NUM, PEDAL_NUMBER);
+
+DrawMenu myMenu(&Wire, SDA0_PIN, SCL0_PIN);
 
 enum MenuState{
   DEDFALUT,
@@ -91,216 +56,6 @@ int currentSelection = 0;
 
 int _test_value = 0;
 
-/*
-Menu Drwaing Functions
-*/
-void display_init(){
-  Wire.setSDA(SDA0_PIN);
-  Wire.setSCL(SCL0_PIN);
-
-  if(!display.begin(SSD1306_SWITCHCAPVCC, SCREEN_ADDRESS)) {
-    Serial.println(F("SSD1306 allocation failed"));
-    for(;;); // Don't proceed, loop forever
-  }
-
-  display.display();
-
-  delay(2000);
-
-  display.clearDisplay();
-  delay(1000);
-
-}
-
-void drawCentreString(const String &buf)
-{
-  int16_t x1, y1;
-  uint16_t w, h;
-  display.getTextBounds(buf, SCREEN_WIDTH/2, SCREEN_HEIGHT/2, &x1, &y1, &w, &h);
-  display.setTextColor(SSD1306_WHITE);
-  display.setCursor(SCREEN_WIDTH/2 - w / 2, SCREEN_HEIGHT/2 - h / 2);
-  display.print(buf);
-}
-
-void drawCentreNumber(const int Number){
-
-  if(Number >= 100 || Number <= -10){
-    display.setTextSize(3);
-  }else{
-    display.setTextSize(4);
-  }
-  drawCentreString(String(Number));
-}
-
-void drawMainMenu(int select){
-  const String MenuString[3] = {"Exit", "Assign", "Setting"};
-  display.clearDisplay();
-  display.setTextSize(2);
-  display.setTextColor(SSD1306_WHITE);
-
-
-  for(int i=0; i<3; i++){
-    display.setCursor(FONT_WIDTH*2, (FONT_HEIGHT*3 + 4)*i);
-    display.print(MenuString[i]);
-  }
-
-
-  display.setCursor(0, (FONT_HEIGHT*3 + 4)*select);
-  display.print(">");
-
-  display.display();
-
-}
-
-
-void drawAssignMenu(int select){
-
-  const String AssignMenuString[3] = {"Exit", "Pedal", "Pedal EXP"};
-  static int displayRenge = 0;
-
-  if(select >= 3 && select >= displayRenge+2){
-    displayRenge = select - 2;
-  }else if(select < displayRenge){
-    displayRenge = select;
-  }
-  
-  display.clearDisplay();
-  display.setTextSize(2);
-  display.setTextColor(SSD1306_WHITE);
-
-
-  for(int i=0; i<3; i++){
-    display.setCursor(FONT_WIDTH*2, (FONT_HEIGHT*3 + 4)*i);
-    
-    if(i+displayRenge == 0){
-      display.print(AssignMenuString[0]);
-    }else if(i+displayRenge == PEDAL_NUMBER + 1){
-      display.print(AssignMenuString[2]);
-    }else if(i+displayRenge < PEDAL_NUMBER+1){
-      display.print(AssignMenuString[1] + " " + String(i+displayRenge));
-    }
-    
-  }
-
-  display.setCursor(0, (FONT_HEIGHT*3 + 4)*(select-displayRenge));
-  display.print(">");
-
-  display.display();
-  Serial.println(select);
-
-}
-
-void drawPedalMenu(int select){
-
-  const String PedalMenuString[3] = {"Exit", "PC", "CC"};
-  display.clearDisplay();
-  display.setTextSize(2);
-  display.setTextColor(SSD1306_WHITE);
-
-
-  for(int i=0; i<3; i++){
-    display.setCursor(FONT_WIDTH*2, (FONT_HEIGHT*3 + 4)*i);
-    display.print(PedalMenuString[i]);
-  }
-
-
-  display.setCursor(0, (FONT_HEIGHT*3 + 4)*select);
-  display.print(">");
-
-  display.display();
-
-}
-
-void drawPCMenu(int select, int Num, int Ch, bool invert){
-  const String PCMenuString[3] = {"Exit", "No:", "CH:"};
-  int Data[2] = {Num, Ch};
-  
-  display.clearDisplay();
-  display.setTextSize(2);
-  display.setTextColor(SSD1306_WHITE);
-
-
-  for(int i=0; i<3; i++){
-    display.setCursor(FONT_WIDTH*2, (FONT_HEIGHT*3 + 4)*i);
-    if(i>0){
-      display.print(PCMenuString[i]);
-      if(invert && select == i)display.setTextColor(SSD1306_BLACK, SSD1306_WHITE);
-      display.print(String(Data[i]));
-      if(invert && select == i)display.setTextColor(SSD1306_WHITE);
-    }else{
-      display.print(PCMenuString[i]);
-    }
-  }
-  display.setCursor(0, (FONT_HEIGHT*3 + 4)*select);
-  display.print(">");
-
-  display.display();
-
-}
-
-void drawCCMenu(int select, int value, int Num, int Ch, bool invert){
-
-  const String CCMenuString[4] = {"Exit", "Value:", "No:", "CH:"};
-  int Data[3] ={value, Num, Ch};
-  static int displayRenge = 0;
-
-  if(select >= 3 && select >= displayRenge+2){
-    displayRenge = select - 2;
-  }else if(select < displayRenge){
-    displayRenge = select;
-  }
-  
-  display.clearDisplay();
-  display.setTextSize(2);
-  display.setTextColor(SSD1306_WHITE);
-
-
-  for(int i=0; i<3; i++){
-    display.setCursor(FONT_WIDTH*2, (FONT_HEIGHT*3 + 4)*i);
-    display.print(CCMenuString[i+displayRenge]);
-    if(displayRenge + i > 0){
-      if(invert && select == i + displayRenge)display.setTextColor(SSD1306_BLACK, SSD1306_WHITE);
-      display.print(String(Data[i + displayRenge]));
-      if(invert && select == i + displayRenge)display.setTextColor(SSD1306_WHITE);
-    }
-  }
-
-  display.setCursor(0, (FONT_HEIGHT*3 + 4)*(select-displayRenge));
-  display.print(">");
-  
-
-  display.display();
-
-}
-
-void drawExpCCMenu(int select, int Num, int Ch, bool invert){
-
-  const String ExpCCMenuString[3] = {"Exit", "No:", "CH:"};
-  int Data[2] = {Num, Ch};
-  
-  display.clearDisplay();
-  display.setTextSize(2);
-  display.setTextColor(SSD1306_WHITE);
-
-
-  for(int i=0; i<3; i++){
-    display.setCursor(FONT_WIDTH*2, (FONT_HEIGHT*3 + 4)*i);
-    if(i>0){
-      display.print(ExpCCMenuString[i]);
-      if(invert && select == i)display.setTextColor(SSD1306_BLACK, SSD1306_WHITE);
-      display.print(String(Data[i]));
-      if(invert && select == i)display.setTextColor(SSD1306_WHITE);
-    }else{
-      display.print(ExpCCMenuString[i]);
-    }
-  }
-  display.setCursor(0, (FONT_HEIGHT*3 + 4)*select);
-  display.print(">");
-
-  display.display();
-
-}
-
 
 /*
 Callback Functions
@@ -322,11 +77,11 @@ void onEncorderChange(){
 
   switch (currentMenuState) {
     case MAIN_MENU:
-      drawMainMenu(currentSelection);
+      myMenu.drawMainMenu(currentSelection);
       break;
 
     case ASSIGN_MENU:
-      drawAssignMenu(currentSelection);
+      myMenu.drawAssignMenu(currentSelection);
       break;
   }
   
@@ -339,7 +94,7 @@ void onSwitchChange(){
 
       }else if(currentSelection == 1){
         currentMenuState = ASSIGN_MENU;
-        drawAssignMenu(currentSelection);
+        myMenu.drawAssignMenu(currentSelection);
         currentSelection = 0;
       }else if(currentSelection == 2){
         
@@ -353,7 +108,7 @@ void setup() {
 
   pixels.begin();
 
-  display_init();
+  myMenu.init(SCREEN_WIDTH, SCREEN_HEIGHT);
 
   myRE.attachRotaryCallback(onEncorderChange);
   myRE.attachSwitchCallback(onSwitchChange);
@@ -369,7 +124,7 @@ void setup1(){
 
   myPedals.begin();
 
-  usb_midi.setStringDescriptor("MyPedal");
+  usb_midi.setStringDescriptor("PiPedal");
 
   MIDI.begin(1);
